Selesai mengatur router all
