+++ src/scripts/routes/url-parser.js
@@ -9,6 +9,9 @@ const UrlParser = {
     const url = window.location.hash.slice(1).toLowerCase();
     return this._urlSplitter(url);
   },
+
+  _urlSplitter(url) {
+    const urlsSplits = url.split("/");
     return {
       resource: urlsSplits[1] || null,
       id: urlsSplits[2] || null,
@@ -17,6 +20,10 @@ const UrlParser = {
   },
 
   _urlCombiner(splitedUrl) {
+    return (
+      (splitedUrl.resource ? `/${splitedUrl.resource}` : "/") +
+      (splitedUrl.id ? "/:id" : "") +
+      (splitedUrl.verb ? `/${splitedUrl.verb}` : "")
     );
   },
 };
