+++ src/styles/index.css
@@ -1,0 +1,156 @@
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+@import "flowbite";
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700;800;900&display=swap');
+
+*{
+  font-family: 'Montserrat', sans-serif;
+}
+
+@keyframes fadeIn {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeOut {
+  0% {
+    transform: translateY(0);
+  }
+  100% {
+    transform: translateY(-100%);
+  }
+}
+
+.animate-fade-out {
+  transition: 2s ease-in-out;
+}
+
+.active {
+  @apply bg-blue-500 text-sky-950 font-bold;
+}
+
+#calendar {
+  width: 100%;
+}
+
+#calendar .fc-day {
+  border: 1px solid #000; /* Warna dan ketebalan border dapat disesuaikan */
+  padding: 5px; /* Tambahkan padding jika diperlukan */
+  box-sizing: border-box; /* Untuk memastikan padding tidak mempengaruhi ukuran total */
+}
+#calendar .fc-day-header {
+  border: 1px solid #000;
+}
+@media screen and (min-width: 600px) {
+  #calendar {
+    width: 90%;
+  }
+  #calendar .fc-toolbar-title,
+  #calendar .fc-button {
+    font-size: 16px;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  #calendar {
+    width: 100%;
+  }
+
+  #calendar .fc-toolbar-title,
+  #calendar .fc-button {
+    font-size: 12px;
+  }
+}
+
+@media screen and (max-width: 450px) {
+  #calendar {
+    width: 100%;
+  }
+
+  #calendar .fc-toolbar-title,
+  #calendar .fc-button {
+    font-size: 10px;
+  }
+}
+
+#kalendar {
+  width: 100%;
+}
+
+#kalendar .fc-day {
+  border: 1px solid #000; /* Warna dan ketebalan border dapat disesuaikan */
+  padding: 5px; /* Tambahkan padding jika diperlukan */
+  box-sizing: border-box; /* Untuk memastikan padding tidak mempengaruhi ukuran total */
+}
+#kalendar .fc-day-header {
+  border: 1px solid #082f49;
+}
+
+@media screen and (min-width: 600px) {
+  #kalendar {
+    width: 90%;
+  }
+  #kalendar .fc-button {
+    font-size: 14px;
+  }
+  #kalendar .fc-toolbar-title {
+    font-size: 24px;
+    font: bold;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  #kalendar {
+    width: 100%;
+  }
+
+  #kalendar .fc-toolbar-title {
+    font-size: 24px;
+    font: bold;
+  }
+
+  #kalendar .fc-button {
+    font-size: 10px;
+  }
+}
+
+@media screen and (max-width: 460px) {
+  #kalendar {
+    width: 100%;
+  }
+  #kalendar .fc-toolbar-title {
+    font-size: 12px;
+    font: bold;
+  }
+  #kalendar .fc-button {
+    font-size: 9px;
+  }
+}
+
+.rotasi {
+  transform: rotate(180deg);
+}
+
+.rotasiP {
+  transform: translateY(-100%);
+}
+
+/* Css untuk Checkbok Notofikasi */
+.on {
+  @apply bg-sky-900;
+}
+
+.blacked-out {
+  background-color: rgba(
+    0,
+    0,
+    0,
+    0.5
+  ); /* 0.5 adalah nilai opasitas, sesuaikan sesuai keinginan Anda */
+  transition: background-color 0.3s ease-in-out;
+}
