+++ src/views/admin/daftar.html
@@ -4,9 +4,12 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <!-- Icons -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"
+    />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" integrity="sha384-H8R4b6PMtUwnNl/b6M7OCjL8aWMIj1hASMAQjDdq23a1VYjTGaK6hJVRbBzKkqI2" crossorigin="anonymous"/>
-    <title>PlanPlan</title>
+    <title>Daftar||PlanPlan</title>
 </head>
 <body>
     <main class="h-screen bg-sky-50">
@@ -74,5 +77,9 @@
           </div>
         </div>
       </main>
+      <script>
+        import '../../styles/index.css';
+        
+      </script>
 </body>
 </html>
