+++ src/views/app.js
@@ -1,4 +1,5 @@
-
+import UrlParser from "../scripts/routes/url-parser";
+import routes from "../scripts/routes/routes";
 class App {
   constructor({ content }) {
     this._content = content;
