+++ src/views/page/dashboard/calendar.js
@@ -67,10 +67,6 @@ const Kalendar = {
               </div>
             </div>
           </div>
-
-
-
-
           <div class="modal w-[70%] z-20 fade bg-white md:w-[70%] lg:w-[40%] pb-12 rounded-2xl absolute hidden" id="modalEvent" tabindex="-1" role="dialog" aria-labelledby="dateModalLabel" aria-hidden="true">
             <span id="closeEvent" class="material-symbols-outlined absolute right-[-5px] bg-red-500 rounded-full text-white cursor-pointer top-[-10px]">close</span>
             <div class="modal-dialog relative border-black" role="document">
