+++ src/views/page/dashboard/indexdash.html
@@ -1,0 +1,145 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+      integrity="sha384-H8R4b6PMtUwnNl/b6M7OCjL8aWMIj1hASMAQjDdq23a1VYjTGaK6hJVRbBzKkqI2"
+      crossorigin="anonymous"
+    />
+
+    <style>
+      .fc-daygrid-day {
+        border: 1px solid #ddd;
+      }
+    </style>
+    <!--  -->
+
+    <title>PlanPlan</title>
+  </head>
+  <body class="bg-sky-50">
+    <header>
+      <nav class="relative w-full border-gray-200 bg-sky-200 dark:bg-gray-900">
+        <div
+          class="flex flex-wrap items-center justify-between max-w-screen-xl p-4 mx-auto"
+        >
+          <a
+            href="/"
+            class="flex items-center space-x-3 rtl:space-x-reverse"
+          >
+            <img src="../logo.jpg" alt="Logo Plan Plan" class="w-14" />
+            <span
+              class="self-center text-2xl font-semibold whitespace-nowrap dark:text-white text-sky-900"
+              >Plan Plan</span
+            >
+          </a>
+          <button
+            data-collapse-toggle="navbar-default"
+            type="button"
+            class="inline-flex items-center justify-center w-10 h-10 p-2 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600"
+            aria-controls="navbar-default"
+            aria-expanded="false"
+          >
+            <span class="sr-only">Open main menu</span>
+            <svg
+              class="w-5 h-5"
+              aria-hidden="true"
+              xmlns="http://www.w3.org/2000/svg"
+              fill="none"
+              viewBox="0 0 17 14"
+            >
+              <path
+                stroke="currentColor"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M1 1h15M1 7h15M1 13h15"
+              />
+            </svg>
+          </button>
+          <div class="hidden w-full md:block md:w-auto" id="navbar-default">
+            <div
+              class="absolute z-50 flex items-center justify-center w-full gap-4 md:relative bg-sky-200 start-0"
+            >
+              <ul
+                class="flex flex-col items-start w-full p-4 mt-4 font-medium md:p-0 md:flex-row md:space-x-8 rtl:space-x-reverse md:mt-0 md:border-0 dark:bg-gray-800 md:dark:bg-gray-900 dark:border-gray-700"
+              >
+                <li>
+                  <a
+                    href="#/dashboard"
+                    class="block px-3 py-2 font-medium rounded menu-item text-sky-900 md:bg-transparent md:p-0 focus:text-white md:focus:text-sky-900"
+                    aria-current="page"
+                    >Beranda</a
+                  >
+                </li>
+                <li>
+                  <a
+                    id="home"
+                    href="#/tentangkami"
+                    class="block px-3 py-2 font-medium rounded menu-item text-sky-900 md:bg-transparent md:p-0 focus:text-white md:focus:text-sky-900"
+                    >Tentang Kami</a
+                  >
+                </li>
+                <li>
+                  <a
+                    id="home"
+                    href="./dashboard.html#/kalender"
+                    class="block px-3 py-2 font-medium rounded menu-item text-sky-900 md:bg-transparent md:p-0 focus:text-white md:focus:text-sky-900"
+                    >Dashboard</a
+                  >
+                </li>
+              </ul>
+              <div
+                class="absolute flex flex-col items-center justify-center cursor-pointer md:px-3 md:scroll-py-10 md:relative"
+              >
+                <div
+                  class="relative flex items-center justify-center w-16 h-16 rounded-full shadow-xl"
+                >
+                  <img
+                    id="profil"
+                    src="../ketua.png"
+                    alt=""
+                    class="bg-cover rounded-full w-14 h-14"
+                  />
+                  <span
+                    id="spanProfil"
+                    class="material-symbols-outlined absolute mt-10 ml-[65%] bg-black text-white rounded-full transition"
+                  >
+                    check_circle
+                  </span>
+                </div>
+                <p
+                  id="nama"
+                  class="absolute hidden mt-20 font-semibold transition w-28 text-sky-900"
+                >
+                  <a href="./profil.html">Ilham Maulana</a>
+                </p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </nav>
+    </header>
+
+    <main class="p-6 contentindexdash"></main>
+    <footer class="bottom-0 w-full shadow bg-sky-200">
+      <div class="w-[95%] border border-t border-t-sky-900 border-sky-200 mx-auto max-w-screen-xl p-4 md:flex md:items-center md:justify-between">
+        <span class="text-sm text-sky-900">© 2023 <a class="hover:underline">PlanPlan™</a>. All Rights Reserved.</span>
+        <ul class="flex flex-wrap items-center mt-3 text-sm font-medium text-gray-500 sm:mt-0">
+          <li>
+            <a href="#/about" class="hover:underline me-4 md:me-6 text-sky-900">Tentang Kami</a>
+          </li>
+          <li>
+            <a href="#/about" class="hover:underline text-sky-900">Kontak</a>
+          </li>
+        </ul>
+      </div>
+    </footer> 
+  </body>
+</html>
