+++ src/views/page/dashboard/todoList.js
@@ -1,3 +1,4 @@
+import '../../../styles/index.css';
 const TodoList = {
   async render() {
     return `
